+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 CHANGELOG
 =========
 
+5.0.0
+-----
+
+* add `StateSerializerInterface` and update `SerializerRegistry`
+* php-xapi/model ^4.0
+* php-xapi/json-test-fixtures ^2.3
+* php-xapi/test-fixtures ^1.3
+* dropped support for PHP < 8.4
+* dropped support for PHPunit < 11.0
+
+
 4.0.0
 -----